Feature/models

+++ src/features/pos.py
@@ -1,5 +1,5 @@
 import pandas as pd
-
+import numpy as np
 from nltk import pos_tag
 
 from .one_hot_encoder import OneHotEncoder
@@ -34,6 +34,6 @@ def apply_pos_one_hot(df: pd.DataFrame):
     ]
     OHE = OneHotEncoder()
     OHE.fit(pos_list)
-    df["pos_categorical"] = df.apply(lambda x: OHE.transform_categorical(x["pos"], x["passage_index"]), axis=1)
+    df["pos_categorical"] = df.apply(lambda x: np.uint8(OHE.transform_categorical(x["pos"], x["passage_index"])), axis=1)
     df["pos_onehot"] = df.apply(lambda x: OHE.transform_one_hot(x["pos_categorical"], x["passage_index"]), axis=1)
     return df, OHE
